pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "delvn"
version = "0.1.0"
description = "Delvn: multi-agent CTI pipeline for correlated threat triage"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2",
  "pydantic-settings",
  "httpx",
  "tenacity",
  "feedparser",
  "python-dateutil",
  "pyyaml",
  "azure-cosmos",
  "azure-search-documents",
  "azure-identity",
  "pytest",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
